remove URL API versioning
In the future when API versioning is needed, we can just do something
like this: https://stripe.com/blog/api-versioning
