+++ config/routes.rb
@@ -1,11 +1,9 @@
 Rails.application.routes.draw do
   if Rails.env.development?
-    mount GraphiQL::Rails::Engine, at: '/graphiql', graphql_path: '/api/v1/graphql'
+    mount GraphiQL::Rails::Engine, at: '/graphiql', graphql_path: '/api/graphql'
   end
 
   namespace :api do
-    namespace :v1 do
-      post 'graphql', to: 'graphql#execute'
-    end
+    post 'graphql', to: 'graphql#execute'
   end
 end
